# packages/backend/openapi-spec.json
{"openapi":"3.0.0","paths":{"/api/event":{"post":{"operationId":"createEvent","summary":"","description":"Create an event","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDto"}}}}},"tags":["event"]},"get":{"operationId":"getAll","summary":"","description":"Get all events","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventDto"}}}}}},"tags":["event"]}},"/api/event/{id}":{"get":{"operationId":"getEvent","summary":"","description":"Get an event","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDto"}}}}},"tags":["event"]}},"/api/user":{"post":{"operationId":"createEvent","summary":"","description":"Create an event","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["user"]},"get":{"operationId":"getAll","summary":"","description":"Get all users","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}},"tags":["user"]}},"/api/user/{address}":{"get":{"operationId":"getUser","summary":"","description":"Get an user","parameters":[{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["user"]}},"/api/file/image":{"post":{"operationId":"uploadFile","summary":"","description":"Upload an image file","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["file"]}}},"info":{"title":"backend","description":"Base backend API","version":"0.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ApiException":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"EventAttributeDto":{"type":"object","properties":{"trait_type":{"type":"string"},"display_type":{"type":"string"},"value":{"type":"object"}},"required":["trait_type","display_type","value"]},"EventDto":{"type":"object","properties":{"id":{"type":"number"},"address":{"type":"string"},"creatorAddress":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"external_url":{"type":"string"},"image":{"type":"string"},"attributes":{"type":"array","items":{"$ref":"#/components/schemas/EventAttributeDto"}}},"required":["id","address","name","description","external_url","image","attributes"]},"UserDto":{"type":"object","properties":{"address":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"image":{"type":"string"}},"required":["address","name","description","image"]}}}}